% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCompare.R
\name{binCompare}
\alias{binCompare}
\title{Graph conditional statistics for a binary outcome and discrete X variable.}
\usage{
binCompare(bt, success, type)
}
\arguments{
\item{bt}{a bivTable object where compare = TRUE}

\item{success}{the level of the outcome you want to graph}

\item{type}{= rrd, rrr, or or}
}
\description{
Graph conditional statistics for a binary outcome and discrete X variable.
}
\examples{
library(carData)
library(soc429Code)
bt <- bivTable(Greene, language, decision, compare = TRUE, baseline = "China")
binCompare(bt, success = "yes", type = "or")
binCompare(bt, success = "yes", type = "rrd")
}
