% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivTable.R
\name{bivTable}
\alias{bivTable}
\title{Create a table of conditional frequencies, relative frequencies, and odds for a single variable}
\usage{
bivTable(data, group, outcome, compare = FALSE, baseline = "")
}
\arguments{
\item{data}{A data frame}

\item{group}{Explanatory or row variable}

\item{outcome}{A dichotomous outcome}

\item{compare}{A logical variable indicating whether or not to compare conditional distributions}

\item{baseline}{A baseline level of the row/explanatory group; required if compare = TRUE}
}
\description{
Create a table of conditional frequencies, relative frequencies, and odds for a single variable
}
\examples{
bivTable(data = carData::GSSvocab, group = ageGroup, outcome = nativeBorn)
}
