% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contCompare.R
\name{contCompare}
\alias{contCompare}
\title{Compare conditional distributions with continuous outcome with optional graph}
\usage{
contCompare(
  data,
  group,
  outcome,
  compare = FALSE,
  baseline = "",
  FUN = "mean",
  quantile = 0.5,
  graph = FALSE,
  graphType = "diff"
)
}
\arguments{
\item{data}{data frame}

\item{group}{your x variable (a factor variable)}

\item{outcome}{your continuous y variable}

\item{compare}{do you want to compare conditional outcomes to a baseline level of x TRUE or FALSE}

\item{baseline}{the omitted/baseline/reference level of x}

\item{FUN}{a function describing an outcome c(mean, median, sd, var, iqr, range)}

\item{quantile}{a vector to pass to quantile()}

\item{graph}{a TRUE/FALSE indicator of whether or not to print a graph}

\item{graphType}{if a graph will be plotted, print differences or ratios}
}
\description{
Compare conditional distributions with continuous outcome with optional graph
}
\examples{
library(car)
data(car)
library(magrittr)
contCompare(data = GSSvocab, group = educGroup, outcome = vocab, graph = FALSE)
contCompare(data = GSSvocab, group = educGroup, outcome = vocab, graph = TRUE,
graphType = "diff", compare = TRUE, baseline = "12 yrs")
contCompare(data = GSSvocab, group = educGroup, outcome = vocab, graph = TRUE,
graphType = "ratio", compare = TRUE, baseline = "16 yrs")
contCompare(data = GSSvocab, group = nativeBorn, outcome = age,
compare = FALSE, graph = TRUE, FUN = "median")
}
