% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durkheimData.R
\name{durkheimData}
\alias{durkheimData}
\title{Generate Data from Durkheim's analysis of punishment & social organization}
\usage{
durkheimData(table = "complexity")
}
\arguments{
\item{table}{simple v. complex outcome or dispersed v. concentrated outcome}
}
\description{
Generate Data from Durkheim's analysis of punishment & social organization
}
\examples{
durkheimData()
}
