% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomMean.R
\name{geomMean}
\alias{geomMean}
\title{Take the geometric mean of a vector of numbers}
\usage{
geomMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numbers}

\item{na.rm}{A logical scalar. Should missing values (including NaN)
be removed? Default is TRUE}
}
\value{
The geometric mean
}
\description{
Take the geometric mean of a vector of numbers
}
\examples{
geomMean(c(1, 2, 3, 4, 5, 6), na.rm = TRUE)
}
