% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBivTable.R
\name{ggBivTable}
\alias{ggBivTable}
\title{Graph a table of relative conditional frequencies as odds or proportions}
\usage{
ggBivTable(table, type = "proportion", failures = 100, success = "")
}
\arguments{
\item{table}{bivTable object. Note that this is NOT a data frame}

\item{type}{"proportion" or "odds"}

\item{failures}{Total number of failures (for odds only)}

\item{success}{The level of the outcome variable that is a success}
}
\description{
Graph a table of relative conditional frequencies as odds or proportions
}
\examples{
bt <- bivTable(data = carData::GSSvocab, group = ageGroup, outcome = nativeBorn)
ggBivTable(table = bt)
ggBivTable(table = bt, type = "odds", success = "no", failures = 20)
}
