% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggUniTable.R, R/ggUniTable_old.R
\name{ggUniTable}
\alias{ggUniTable}
\title{Graph a table of relative marginal frequencies as odds or proportions}
\usage{
ggUniTable(ut, type = "proportion", total = 100, failures = 100, success = "")

ggUniTable(ut, type = "proportion", total = 100, failures = 100, success = "")
}
\arguments{
\item{ut}{uniTable object. Note that this is NOT a data frame}

\item{type}{"proportion" or "odds"}

\item{total}{Total number of observations (for proportions only)}

\item{failures}{Total number of failures (for odds only)}

\item{success}{The level of the outcome variable that is a success}
}
\description{
Graph a table of relative marginal frequencies as odds or proportions

Graph a table of relative marginal frequencies as odds or proportions
}
\examples{
ut <- uniTable(data = carData::Arrests, variable = employed)
ggUniTable(ut = ut)
ggUniTable(ut = ut, type = "odds", success = "No", failures = 10)
ut <- uniTable(data = carData::Arrests, variable = employed)
ggUniTable(ut = ut, total = 600)
ggUniTable(ut = ut, type = "odds", success = "No", failures = 42)
}
