% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggUniTable_old.R
\name{ggUniTable_old}
\alias{ggUniTable_old}
\title{Graph a table of relative marginal frequencies as odds or proportions}
\usage{
ggUniTable_old(
  ut,
  type = "proportion",
  total = 100,
  failures = 100,
  success = ""
)
}
\arguments{
\item{ut}{uniTable object. Note that this is NOT a data frame}

\item{type}{"proportion" or "odds"}

\item{total}{Total number of observations (for proportions only)}

\item{failures}{Total number of failures (for odds only)}

\item{success}{The level of the outcome variable that is a success}
}
\description{
Graph a table of relative marginal frequencies as odds or proportions
}
\examples{
ut <- uniTable(data = carData::Arrests, variable = employed)
ggUniTable_old(ut = ut, total = 600)
ggUniTable_old(ut = ut, type = "odds", success = "No", failures = 42)
}
