% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letMeEatCake.R
\name{letMeEatCake}
\alias{letMeEatCake}
\title{Visualizing the idea of maximum likelihood with a silly cake example}
\usage{
letMeEatCake()
}
\description{
Visualizing the idea of maximum likelihood with a silly cake example
}
\examples{
letMeEatCake()
}
