% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPlot.R
\name{logPlot}
\alias{logPlot}
\title{Graph an outcome variable that has been log transformed}
\usage{
logPlot(
  xstart = 0,
  xincrement = 1,
  int = 0,
  b1 = 1,
  xlog = FALSE,
  ylog = TRUE,
  xmin,
  xmax,
  ymin = NULL,
  ymax = NULL,
  digits = 2
)
}
\arguments{
\item{xstart}{starting value of x}

\item{xincrement}{how much x increases}

\item{int}{intercept}

\item{b1}{slope}

\item{xlog}{whether x is logged or not}

\item{ylog}{whether y is logged or not}

\item{xmin}{minimum of x axis}

\item{xmax}{maximum of x axis}

\item{ymin}{minimum of y axis}

\item{ymax}{maximum of y axis}

\item{digits}{number of digits to print}
}
\description{
Graph an outcome variable that has been log transformed
}
\examples{
library(carData)
m <- lm(data = USArrests, log(Rape) ~ Assault)
texreg::screenreg(m, digits = 3)
int <- m$coefficients["(Intercept)"]
b1 <- m$coefficients["Assault"]
logPlot(xstart = 150, xincrement = 1, int = int, b1 = b1, ylog = TRUE, xmin = 148, xmax = 152)
logPlot(xstart = 148, xincrement = 1, int = int, b1 = b1, ylog = TRUE, xmin = 140, xmax = 152)
logPlot(xstart = 148, xincrement = 100, int = int, b1 = b1, ylog = TRUE, xmin = 0, xmax = 400)
}
