% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMax.R
\name{minMax}
\alias{minMax}
\title{Find the minimum or maximum of a quadratic function of x in a linear model}
\usage{
minMax(M, variable = "")
}
\arguments{
\item{M}{a linear model object created with lm}

\item{variable}{The name of the variable as a string}
}
\description{
Find the minimum or maximum of a quadratic function of x in a linear model
}
\examples{
minMax(M = lm(data = mtcars, mpg ~ wt + I(wt^2)), variable = "wt")
}
