% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycisB.R
\name{mycisB}
\alias{mycisB}
\title{Graph simulated sampling distribution of sample slopes & intercepts}
\usage{
mycisB(N = 5, R = 25, mean = 100, std = 15, intc = -2, slope = 0.25)
}
\arguments{
\item{N}{Sample size}

\item{R}{Number of draws}

\item{mean}{Mean of population}

\item{std}{Standard deviation of population}

\item{intc}{Intercept value}

\item{slope}{Slope value}
}
\description{
Graph simulated sampling distribution of sample slopes & intercepts
}
\examples{
g <- mycisB()
g$g.int
g$g.slope
}
