% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycisMean.R
\name{mycisMean}
\alias{mycisMean}
\title{Simulate sampling distributions of sample means}
\usage{
mycisMean(N = 5, R = 10, mean = 100, std = 15)
}
\arguments{
\item{N}{Sample size}

\item{R}{Number of draws}

\item{mean}{Mean of population}

\item{std}{Standard deviation of population}
}
\description{
Simulate sampling distributions of sample means
}
\examples{
mycisMean()
mycisMean(N = 100, R = 20)
}
