% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycisVar.R
\name{mycisVar}
\alias{mycisVar}
\title{Simulate normal distributions for the purpose of illustrating desirable properties of estimators}
\usage{
mycisVar(N = 5, R = 10, mean = 100, std = 15)
}
\arguments{
\item{N}{Number of samples}

\item{R}{Number of replications}

\item{mean}{Mean of population}

\item{std}{Standard deviation of population}
}
\description{
Simulate normal distributions for the purpose of illustrating desirable properties of estimators
}
\examples{
p1 <- mycisVar(N = 10, R = 10)
g1 <- p1$g + ggplot2::scale_x_continuous(limits = c(0, 1700))
p2 <- mycisVar(N = 50, R = 10)
g2 <- p2$g + ggplot2::scale_x_continuous(limits = c(0, 1700))
p3 <- mycisVar(N = 500, R = 10)
g3 <- p3$g + ggplot2::scale_x_continuous(limits = c(0, 1700))
gridExtra::grid.arrange(g1, g2, g3, nrow = 3)
}
