% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myx.R
\name{myx}
\alias{myx}
\title{Simulate normal distributions for the purpose of illustrating desirable properties of estimators}
\usage{
myx(
  N = 5,
  mean = 100,
  std = 15,
  seed = 12345,
  biased = FALSE,
  re = TRUE,
  ic = FALSE
)
}
\arguments{
\item{N}{Number of samples}

\item{mean}{Mean of population}

\item{std}{Standard deviation of population}

\item{seed}{Random number seed}

\item{biased}{Generate a biased sampling distribution? True or False}

\item{re}{Generate a relatively efficient sampling distribution? True or False}

\item{ic}{Generate an inconsistent sampling distribution? True or False}
}
\description{
Simulate normal distributions for the purpose of illustrating desirable properties of estimators
}
\examples{
g <- myx(N = 10, biased = TRUE)
suppressWarnings(print(g$g.pop))
g$g.sampdist
myx(biased = FALSE, re = TRUE)$g.sampdist +
ggplot2::ggtitle("Relatively Efficient (N = 10)")
myx(biased = FALSE, re = FALSE)$g.sampdist +
ggplot2::ggtitle("Relatively Inefficient (N = 10)")
g <- myx(N = 100, biased = FALSE)
g$g.pop
myx(N = 5, biased = TRUE)$g.sampdist +
ggplot2::ggtitle("Cindy: Narrower, but biased so not consistent", "Small N")
myx(N = 100, biased = TRUE)$g.sampdist +
ggplot2::ggtitle("Cindy: Narrower, but biased so not consistent", "Large N")
}
