% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orSE.R
\name{orSE}
\alias{orSE}
\title{Calculate the standard error for an odds ratio by exponentiating the log}
\usage{
orSE(model)
}
\arguments{
\item{model}{a GLM model with family = "binomial"}
}
\description{
Calculate the standard error for an odds ratio by exponentiating the log
}
\examples{
m <- glm(data = mtcars, factor(am) ~ mpg + disp, family = "binomial")
summary(m)
orSE(m)
}
