% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition a data frame into two pieces}
\usage{
partition(data = df, seed = 12345, tprop = 0.2)
}
\arguments{
\item{data}{A data frame}

\item{seed}{Random number seed}

\item{tprop}{Proportion for training data (between 0 and 1)}
}
\description{
Partition a data frame into two pieces
}
\examples{
partition(mtcars, seed = 99999, tprop = 0.5)
}
