% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printHelp.R
\name{printHelp}
\alias{printHelp}
\title{Print help file associated with a function}
\usage{
printHelp(topic, format = c("text", "html", "latex", "Rd"), lines = NULL)
}
\arguments{
\item{topic}{Name of function}

\item{format}{Optional format}

\item{lines}{Optional lines to pull out}
}
\value{
Printed output
}
\description{
Print help file associated with a function
}
\examples{
printHelp(USArrests)
}
