% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_Fixed.R
\name{rnorm_fixed}
\alias{rnorm_fixed}
\title{Create a normal random variable of a given length with fixed mean and standard deviation}
\usage{
rnorm_fixed(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{Number of observations}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\description{
Create a normal random variable of a given length with fixed mean and standard deviation
}
\examples{
rnorm_fixed(100, mu = 0, sigma = 1)
}
