% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlot.R
\name{simplePlot}
\alias{simplePlot}
\title{simplePlot}
\usage{
simplePlot(
  data,
  xcol,
  fill = "white",
  color = "black",
  bins = 5,
  H = TRUE,
  D = TRUE,
  B = TRUE,
  P = TRUE
)
}
\arguments{
\item{data}{The data frame that the variable is in}

\item{xcol}{The variable of interest, entered as a string}

\item{fill}{The fill color of the plots, entered as a string}

\item{color}{The outline color of the plots, entered as a string}

\item{bins}{The number of bins for the plots, entered as a numeric value}

\item{H}{Boolean operator to create histogram, default is FALSE}

\item{D}{Boolean operator to create density plot, default is FALSE}

\item{B}{Boolean operator to create boxplot, default is FALSE}

\item{P}{Boolean operator to create freqpoly plot, default is FALSE}
}
\value{
The requested plots in ggplot2 format
}
\description{
Generate plots for a continuous/quantitative univariate variable using the ggplot2 package
}
\examples{
simplePlot(mtcars, "wt")
simplePlot(mtcars, "wt", H = FALSE, B = FALSE, P = FALSE, color = "purple", fill = "pink")
}
