% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizedExample429.R
\name{standardizedExample429}
\alias{standardizedExample429}
\title{Regression with standardized variables}
\usage{
standardizedExample429(scale = FALSE)
}
\arguments{
\item{scale}{Whether variables are standardized or not}
}
\description{
Regression with standardized variables
}
\examples{
standardizedExample429()
}
