% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniTable.R
\name{uniTable}
\alias{uniTable}
\title{Create a table of frequencies, relative frequencies, and odds for a single variable}
\usage{
uniTable(data, variable)
}
\arguments{
\item{data}{A data frame}

\item{variable}{A variable}
}
\description{
Create a table of frequencies, relative frequencies, and odds for a single variable
}
\examples{
library(carData)
uniTable(data = carData::Arrests, variable = employed)
}
